scripts/species/training/requirements.txt
sklearn
keras
tensorflow
jupyterlab
pandas
google-cloud-storage
pillow
tqdm
tfds-nightly